:bug: Fix (~Tx) in PGN notation
Before this commit, the pgnFuncs.fromMove(...) used to append (~Tx)
anytime a new timeline is created (the set difference between the new
and old sets of active timelines did not take into account the newly
created timeline) and the value for `x` would only be its internal
index.

With this fix, (~Tx) will only be shown if:
- an inactive timeline is re-activated
- that inactive timeline shifts the present further back (if the present
  is only shifted by a half-turn further back, (~Tx) will *not* be
shown)

The value for `x` will then be that of the turn coordinate of the last
board of the newly activated timeline.

See https://discord.com/channels/715057032901951509/715057033514188833/818110950753566741